requirements.txt
asgiref==3.8.1
Django==5.0.4
django-ckeditor==6.7.1
django-js-asset==2.2.0
django-nested-admin==4.0.2
django-nested-inline==0.4.6
pillow==10.3.0
python-monkey-business==1.0.0
six==1.16.0
sqlparse==0.5.0
transliterate==1.10.2
